support all filters, separate tests

+++ Libs/GoBearCore/extended/gbslider.py
@@ -30,7 +30,7 @@ class GBSlider(ExtWebElement):
         ExtWebElement.__init__(self, we)
 
     def get_max_value(self):
-        return int(self.get_element_contains_class(SLIDER_SLIDER_MAX_VALUE).get_attribute(SLIDER_SLIDER_VALUE_ATTRIBUTE))
+        return float(self.get_element_contains_class(SLIDER_SLIDER_MAX_VALUE).get_attribute(SLIDER_SLIDER_VALUE_ATTRIBUTE))
 
     def get_min_handle(self):
         return self.get_element_contains_class(SLIDER_SLIDER_HANDLE_MIN)
@@ -46,10 +46,10 @@ class GBSlider(ExtWebElement):
 
     def set_min_value(self, value):
         action = ActionChains(self.parent)
-        distance = int(value)*(self.get_selection_bar_width())/self.get_max_value()
+        distance = float(value)*(self.get_selection_bar_width())/self.get_max_value()
         action.click_and_hold(self.get_min_handle()).move_by_offset(distance, 0).release().perform()
 
     def set_max_value(self, value):
         action = ActionChains(self.parent)
-        distance = int(value)*(self.get_selection_bar_width())/self.get_max_value()
+        distance = float(value)*(self.get_selection_bar_width())/self.get_max_value()
         action.click_and_hold(self.get_max_handle()).move_by_offset(-distance, 0).release().perform()
